+++ src/dmote_keycap/core.clj
@@ -71,6 +71,9 @@
    [nil "--skirt-thickness N" "Thickness of walls descending around switch"
     :default-desc (str (:skirt-thickness data/option-defaults))
     :parse-fn #(Float/parseFloat %)]
+   [nil "--skirt-space N" "Gap between switch and skirt"
+    :default-desc (str (:skirt-gap data/option-defaults))
+    :parse-fn #(Float/parseFloat %)]
    [nil "--nozzle-width N" "FDM printer nozzle (aperture) width in mm"
     :default-desc (str (:nozzle-width data/option-defaults))
     :parse-fn #(Float/parseFloat %)]