+++ src/dmote_keycap/data.clj
@@ -74,7 +74,7 @@
                                           :y "1"}}]))}
    :nozzle-width 0.5
    :horizontal-support-height 0.5
-   :error-body-positive -0.5
+   :error-body-positive 0
    :error-side-negative 0
    :error-stem-negative 0
    :error-stem-positive 0